refactor code and updated plots
