+++ biorxiv/pca_association_experiment/biorxiv_pca_plots.py
@@ -10,6 +10,8 @@
 
 get_ipython().run_line_magic('load_ext', 'autoreload')
 get_ipython().run_line_magic('autoreload', '2')
+import sys
+sys.path.append("../../modules/")
 
 import pandas as pd
 from scipy.spatial.distance import cdist
